Move windows to windows/ dir
