+++ rtgui/editorpanel.cc
@@ -28,7 +28,6 @@
 #include "batchqueueentry.h"
 #include "soundman.h"
 #include "rtimage.h"
-#include "rtwindow.h"
 #include "filepanel.h"
 #include "guiutils.h"
 #include "popupbutton.h"
@@ -39,9 +38,10 @@
 #include "procparamchangers.h"
 #include "placesbrowser.h"
 #include "pathutils.h"
-#include "rtappchooserdialog.h"
 #include "thumbnail.h"
 #include "toolpanelcoord.h"
+#include "windows/rtappchooserdialog.h"
+#include "windows/rtwindow.h"
 
 #ifdef _WIN32
 #include "windows.h"
