+++ rtgui/filecatalog.cc
@@ -29,22 +29,22 @@
 #include "rtengine/rt_math.h"
 #include "rtengine/procparams.h"
 
-#include "guiutils.h"
-#include "options.h"
-#include "rtimage.h"
-#include "cachemanager.h"
-#include "multilangmgr.h"
-#include "coarsepanel.h"
-#include "filepanel.h"
-#include "renamedlg.h"
-#include "thumbimageupdater.h"
 #include "batchqueue.h"
 #include "batchqueueentry.h"
-#include "placesbrowser.h"
+#include "cachemanager.h"
+#include "coarsepanel.h"
+#include "filepanel.h"
+#include "guiutils.h"
+#include "inspector.h"
+#include "multilangmgr.h"
+#include "options.h"
 #include "pathutils.h"
+#include "placesbrowser.h"
+#include "rtimage.h"
+#include "thumbimageupdater.h"
 #include "thumbnail.h"
 #include "toolbar.h"
-#include "inspector.h"
+#include "windows/renamedlg.h"
 
 using namespace std;
