+++ rtgui/windows/editwindow.cc
