+++ rtgui/windows/partialpastedlg.cc
