+++ rtgui/windows/partialpastedlg.h
