+++ rtgui/windows/renamedlg.cc
