+++ rtgui/windows/renamedlg.h
