+++ rtgui/windows/splash.cc
