+++ rtgui/windows/splash.h
